Fix an inefficiency in uv-sync in which the uv-catalog is always sent even if the uv-hash matches.  Also add preliminary change to support uv-push.
FossilOrigin-Name: d015a269e63fe4c34bc16e41cd26695850e8a196381351d8494298d9f3f36d66

+++ src/blob.c
@@ -462,7 +462,7 @@ int blob_constant_time_cmp(Blob *pA, Blob *pB){
 int blob_eq_str(Blob *pBlob, const char *z, int n){
   Blob t;
   blob_is_init(pBlob);
-  if( n<=0 ) n = strlen(z);
+  if( n<=0 ) n = (int)strlen(z);
   t.aData = (char*)z;
   t.nUsed = n;
   t.xRealloc = blobReallocStatic;
